lightningd/closing_control: routines to control closingd (move from peer_control.c)
